CTCP-4697: Move NetWeight controller to HC directory
