+++ app/pages/houseConsignment/index/items/NetWeightPage.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package pages
+package pages.houseConsignment.index.items
 
 import models.Index
+import pages.QuestionPage
 import pages.sections.ItemsSection
 import play.api.libs.json.JsPath
