+++ conf/app.houseConsignment.index.items.routes
@@ -41,6 +41,11 @@ POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex
 GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/change-gross-weight                    controllers.houseConsignment.index.items.GrossWeightController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = CheckMode)
 POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/change-gross-weight                    controllers.houseConsignment.index.items.GrossWeightController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = CheckMode)
 
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/net-weight                          controllers.houseConsignment.index.items.NetWeightController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = NormalMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/net-weight                          controllers.houseConsignment.index.items.NetWeightController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = NormalMode)
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/change-net-weight                   controllers.houseConsignment.index.items.NetWeightController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = CheckMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/change-net-weight                   controllers.houseConsignment.index.items.NetWeightController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = CheckMode)
+
 GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/commodity-code/remove                  controllers.houseConsignment.index.items.RemoveCommodityCodeYesNoController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = NormalMode)
 POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/commodity-code/remove                  controllers.houseConsignment.index.items.RemoveCommodityCodeYesNoController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = NormalMode)
 GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/change-commodity-code/remove           controllers.houseConsignment.index.items.RemoveCommodityCodeYesNoController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = CheckMode)
