+++ test/controllers/documents/AddAdditionalInformationYesNoControllerSpec.scala
@@ -21,7 +21,6 @@ import forms.YesNoFormProvider
 import models.NormalMode
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
-import pages.AddUnloadingCommentsYesNoPage
 import pages.documents.AddAdditionalInformationYesNoPage
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.test.FakeRequest
